Move Scheduler from A2C utils to common schedules file
